Usage

To use the DividedRectangles module, start your code with:

using DividedRectangles

The optimize function is the primary function of the DividedRectangles module. It implements the DIRECT algorithm to find the minimum of a given objective function within specified bounds.

To use the optimize function with a custom objective function::

using DividedRectangles

# Define the objective function
f(x) = x[1]^2 + x[2]^2 + 3 * sin(5 * x[1]) + 2 * cos(3 * x[2])  # Multivariate example

# Set the search bounds
a = [-2.0, -2.0]
b = [2.0, 2.0]

# Call the optimization function
result = optimize(f, a, b)

println("Best design found: ", result)

Arguments:

f: The objective function to be minimized.

a: Vector of lower bounds for the search space.

b: Vector of upper bounds for the search space.

max_iterations: (Optional) The maximum number of iterations (default: 100).

min_radius: (Optional) The minimum radius of hyper-rectangles (default: 1e-5).

Returns:

The best design x found by DIRECT.

The package also provides direct, which is the same as optimize except it returns all hyperrectangular intervals:

using DividedRectangles

# Define the objective function
f(x) = x[1]^2 + x[2]^2 + 3 * sin(5 * x[1]) + 2 * cos(3 * x[2])  # Multivariate example

# Set the search bounds
a = [-2.0, -2.0]
b = [2.0, 2.0]

# Run DIRECT
intervals = direct(f, a, b, max_iterations=10, min_radius=1e-4)

Arguments:

f: The objective function to be minimized.

a: Vector of lower bounds for the search space.

b: Vector of upper bounds for the search space.

max_iterations: (Optional) The maximum number of iterations (default: 100).

min_radius: (Optional) The minimum radius of hyper-rectangles (default: 1e-5).

Returns:

All hyperrectangular intervals maintained by DIRECT.

Functions and Types

DividedRectangles.basis — Function

A helper function that returns a basis vector with a single 1 entry in an otherwise zero vector.

DividedRectangles.is_ccw — Function

A helper function that determines whether a→b→c is counter-clockwise in (r,y) space.

DividedRectangles.DirectRectangle — Type

A structure for store information about each hyperrectangular interval.

DividedRectangles.split_interval — Function

Split the given interval, where g is the objective function in the unit hypercube. This method returns a list of the resulting smaller intervals.

DividedRectangles.get_split_intervals — Function

A routine for obtaining the split intervals from a given list of intervals and a minimum radius. The potentially optimal intervals form a lower-right convex hull in r and y.

DividedRectangles.direct — Function

An implementation of DIRECT that runs for the given number of iterations and then returns all hyperrectangular intervals.

DividedRectangles.optimize — Function

The primary method provided by DividedRectangles.jl, which is used to optimize an objective function and return the best design found.